Re-add rubyBlockParameter
For somewhat arbitrary reasons it seems that rubyBlockParameter was removed from
vim-ruby's syntax/ruby.vim. Re-add it so that we have something to link the new
rubyNumberedBlockParam to, and also because I vaguely remember having it and it
was nice.

Also introduce rubyNumberedBlockParamError to help Rubyists avoid using `_0` by
mistake.

https://github.com/vim-ruby/vim-ruby/commit/074200ffa39b19baf9d9750d399d53d97f21ee07
https://github.com/vim-ruby/vim-ruby/commit/eba523334fe14e5c9b9585a287a336317da26d70